requirements.txt
streamlit
openai
python-dotenv
PyPDF2
python-docx
requests 
beautifulsoup4 
googlesearch-python 
fpdf
google
selenium  
pdfkit
reportlab
langchain
langchain-community 
langchainhub 
langchain-openai 
langchain-chroma 
bs4
faiss-cpu
pypdf